+++ js/classes/CalendarFactory.js
@@ -5,6 +5,7 @@ const { getDefaultWidthHeight} = require('../user-preferences.js');
 const { Calendar } = require('./Calendar.js');
 const { FixedDayCalendar } = require('./FixedDayCalendar.js');
 const { FlexibleMonthCalendar } = require('./FlexibleMonthCalendar.js');
+const { FlexibleDayCalendar } = require('./FlexibleDayCalendar.js');
 
 class CalendarFactory {
     static getInstance(preferences, calendar = undefined) {
@@ -38,7 +39,18 @@ class CalendarFactory {
             throw new Error(`Could not instantiate ${view}`);
         }
         else if (numberOfEntries === 'flexible') {
-            if (view === 'month') {
+            if (view === 'day') {
+                if (calendar === undefined || calendar.constructor.name !== 'FlexibleDayCalendar') {
+                    if (calendar !== undefined && calendar.constructor.name !== 'FlexibleDayCalendar') {
+                        ipcRenderer.send('RESIZE_MAIN_WINDOW', widthHeight.width, widthHeight.height);
+                    }
+                    return new FlexibleDayCalendar(preferences);
+                } else {
+                    calendar.updatePreferences(preferences);
+                    calendar.redraw();
+                    return calendar;
+                }
+            } else if (view === 'month') {
                 if (calendar === undefined || calendar.constructor.name !== 'FlexibleMonthCalendar') {
                     if (calendar !== undefined && calendar.constructor.name !== 'FlexibleMonthCalendar') {
                         ipcRenderer.send('RESIZE_MAIN_WINDOW', widthHeight.width, widthHeight.height);