+++ js/classes/FixedDayCalendar.js
@@ -8,6 +8,7 @@ const {
 } = require('../time-math.js');
 const { getDateStr, getMonthLength } = require('../date-aux.js');
 const { Calendar } = require('./Calendar.js');
+const { generateKey } = require('../date-db-formatter');
 
 class FixedDayCalendar extends Calendar {
     /**
@@ -39,7 +40,7 @@ class FixedDayCalendar extends Calendar {
      * Generates the calendar HTML view.
      */
     _generateTemplate() {
-        var body = this._getBody();
+        let body = this._getBody();
         $('#calendar').html(body);
         $('html').attr('data-view', 'day');
     }
@@ -70,7 +71,7 @@ class FixedDayCalendar extends Calendar {
      * Returns the template code of the body of the page.
      */
     _getBody() {
-        var html = '<div>';
+        let html = '<div>';
         html += this.constructor._getPageHeader();
         html += '<div id="calendar-table-body">';
         html += '</div>';
@@ -127,7 +128,7 @@ class FixedDayCalendar extends Calendar {
     _getInputsRowCode(year, month, day) {
         let today = new Date(),
             isToday = (today.getDate() === day && today.getMonth() === month && today.getFullYear() === year),
-            trID = ('tr-' + year + '-' + month + '-' + day);
+            trID = ('tr-' + generateKey(year, month, day));
 
         if (!this._showDay(year, month, day)) {
             return '<div class="today-non-working" id="' + trID + '">' +