+++ js/menus.js
@@ -11,7 +11,7 @@ const {
 } = require('./saved-preferences.js');
 const { importDatabaseFromFile, exportDatabaseToFile, migrateFixedDbToFlexible } = require('./import-export.js');
 const { notify } = require('./notification');
-const { os } = require('os');
+const os = require('os');
 const { savePreferences } = require('./user-preferences.js');
 const path = require('path');
 const Store = require('electron-store');
@@ -106,7 +106,7 @@ function getContextMenuTemplate(mainWindow)
         {
             label: 'Punch time', click: function()
             {
-                var now = new Date();
+                let now = new Date();
 
                 mainWindow.webContents.executeJavaScript('calendar.punchDate()');
                 // Slice keeps "HH:MM" part of "HH:MM:SS GMT+HHMM (GMT+HH:MM)" time string
@@ -134,7 +134,7 @@ function getDockMenuTemplate(mainWindow)
         {
             label: 'Punch time', click: function()
             {
-                var now = new Date();
+                let now = new Date();
 
                 mainWindow.webContents.executeJavaScript('calendar.punchDate()');
                 // Slice keeps "HH:MM" part of "HH:MM:SS GMT+HHMM (GMT+HH:MM)" time string
@@ -180,7 +180,7 @@ function getEditMenuTemplate(mainWindow)
                 }
 
                 const htmlPath = path.join('file://', __dirname, '../src/preferences.html');
-                prefWindow = new BrowserWindow({ width: 400,
+                prefWindow = new BrowserWindow({ width: 450,
                     height: 560,
                     parent: mainWindow,
                     resizable: true,
@@ -244,7 +244,7 @@ function getEditMenuTemplate(mainWindow)
             label: 'Export database',
             click()
             {
-                var options = {
+                let options = {
                     title: 'Export DB to file',
                     defaultPath : 'time_to_leave',
                     buttonLabel : 'Export',
@@ -273,7 +273,7 @@ function getEditMenuTemplate(mainWindow)
             label: 'Import database',
             click()
             {
-                var options = {
+                let options = {
                     title: 'Import DB from file',
                     buttonLabel : 'Import',