+++ js/time-math.js
@@ -4,8 +4,8 @@
  * Formats hour, min into string HH:MM
  */
 function hourMinToHourFormatted(hours, minutes) {
-    var paddingHour = hours < 10 ?  '0' : '';
-    var paddingMin = minutes < 10 ?  '0' : '';
+    let paddingHour = hours < 10 ?  '0' : '';
+    let paddingMin = minutes < 10 ?  '0' : '';
     return paddingHour + hours +
            ':' +
            paddingMin + minutes;
@@ -23,11 +23,11 @@ function isNegative(str) {
  * Hours must be formated as HH:MM
  */
 function hourToMinutes(time) {
-    var st = time.split(':');
-    var isNeg = isNegative(time);
+    let st = time.split(':');
+    let isNeg = isNegative(time);
     st[0] = isNeg ? st[0].substr(1) : st[0];
 
-    var min = Number(st[1]) + (Number(st[0]) * 60);
+    let min = Number(st[1]) + (Number(st[0]) * 60);
     if (isNeg) {
         min = min * -1;
     }
@@ -38,12 +38,12 @@ function hourToMinutes(time) {
  * Formats a given amount of minutes into string HH:MM
  */
 function minutesToHourFormatted(min) {
-    var signStr = min < 0 ? '-' : '';
+    let signStr = min < 0 ? '-' : '';
     if (min < 0) {
         min = Math.abs(min);
     }
-    var hours = Math.floor(min / 60);
-    var minutes = Math.floor(min - (hours * 60));
+    let hours = Math.floor(min / 60);
+    let minutes = Math.floor(min - (hours * 60));
     return signStr + hourMinToHourFormatted(hours, minutes);
 }
 
@@ -52,7 +52,7 @@ function minutesToHourFormatted(min) {
  * Time should be formated as HH:MM
  */
 function subtractTime(t1, t2) {
-    var diffMin = hourToMinutes(t2) - hourToMinutes(t1);
+    let diffMin = hourToMinutes(t2) - hourToMinutes(t1);
     return minutesToHourFormatted(diffMin);
 }
 
@@ -61,7 +61,7 @@ function subtractTime(t1, t2) {
  * Time should be formated as HH:MM
  */
 function multiplyTime(t, n) {
-    var totalMin = hourToMinutes(t);
+    let totalMin = hourToMinutes(t);
     totalMin = totalMin * n;
     return minutesToHourFormatted(totalMin);
 }
@@ -71,7 +71,7 @@ function multiplyTime(t, n) {
  * Time should be formated as HH:MM
  */
 function sumTime(t1, t2) {
-    var sumMin = hourToMinutes(t2) + hourToMinutes(t1);
+    let sumMin = hourToMinutes(t2) + hourToMinutes(t1);
     return minutesToHourFormatted(sumMin);
 }
 
@@ -80,7 +80,7 @@ function sumTime(t1, t2) {
  * @returns true if it's valid
  */
 function validateTime(time) {
-    var re = new RegExp('[0-2][0-9]:[0-5][0-9]');
+    let re = new RegExp('[0-2][0-9]:[0-5][0-9]');
     return re.test(time);
 }