+++ js/user-preferences.js
@@ -16,7 +16,7 @@ const defaultPreferences = {
     'repetition': true,
     'notifications-interval': '5',
     'start-at-login': false,
-    'theme': 'light',
+    'theme': 'system-default',
     'overall-balance-start-date' : '2019-01-01',
     'update-remind-me-after' : '2019-01-01',
     'working-days-monday': true,
@@ -55,7 +55,7 @@ function savePreferences(preferencesOptions) {
  * @return {Object}
  */
 function readPreferences() {
-    var preferences;
+    let preferences;
     try {
         preferences = JSON.parse(fs.readFileSync(getPreferencesFilePath()));
     } catch (err) {
@@ -65,7 +65,7 @@ function readPreferences() {
 }
 
 function getDerivedPrefsFromLoadedPrefs(loadedPreferences) {
-    var derivedPreferences = {};
+    let derivedPreferences = {};
     Object.keys(defaultPreferences).forEach(function(key) {
         derivedPreferences[key] = (typeof loadedPreferences[key] !== 'undefined') ? loadedPreferences[key] : defaultPreferences[key];
     });
@@ -83,7 +83,7 @@ function initPreferencesFileIfNotExistsOrInvalid() {
         return;
     }
 
-    var shouldSaveDerivedPrefs = false,
+    let shouldSaveDerivedPrefs = false,
         loadedPrefs = readPreferences(),
         derivedPrefs = getDerivedPrefsFromLoadedPrefs(loadedPrefs),
         loadedPref = Object.keys(loadedPrefs).sort(),
@@ -95,8 +95,8 @@ function initPreferencesFileIfNotExistsOrInvalid() {
     }
 
     // Validate the values
-    for (var key of derivedPrefsKeys) {
-        var value = derivedPrefs[key];
+    for (let key of derivedPrefsKeys) {
+        let value = derivedPrefs[key];
         switch (key) {
         // Handle Time Inputs
         case 'notifications-interval':
@@ -138,12 +138,7 @@ function initPreferencesFileIfNotExistsOrInvalid() {
             shouldSaveDerivedPrefs |= !isValidTheme(value);
             break;
         case 'view':
-            if (derivedPrefs['number-of-entries'] === 'flexible') { // flexible only working with month calendar yet
-                shouldSaveDerivedPrefs |= !(value === 'month');
-            }
-            else {
-                shouldSaveDerivedPrefs |= !(value === 'month' || value === 'day');
-            }
+            shouldSaveDerivedPrefs |= !(value === 'month' || value === 'day');
             break;
         case 'number-of-entries':
             shouldSaveDerivedPrefs |= !(value === 'fixed' || value === 'flexible');
@@ -188,7 +183,7 @@ function showWeekDay(weekDay, preferences = undefined) {
  * @note: The month should be 0-based (i.e.: 0 is Jan, 11 is Dec).
  */
 function showDay(year, month, day, preferences = undefined)  {
-    var currentDay = new Date(year, month, day), weekDay = currentDay.getDay();
+    let currentDay = new Date(year, month, day), weekDay = currentDay.getDay();
     return showWeekDay(weekDay, preferences);
 }